requirements.txt
PyYAML>=6.0
jsonschema>=4.24.0
